pyproject.toml
[project]
name = "shop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django>=5.2.3",
    "django-allauth>=65.9.0",
    "django-weasyprint>=2.4.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "psycopg2-binary>=2.9.10",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ruff>=0.12.1",
    "weasyprint>=65.1",
]
